Checking if logged in user is owner of the image before allowing edit and delete operations.

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -105,13 +105,32 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
+
+        /** PartA Issue #2: https://github.com/upgrad-edu/Course_4_Assignment/issues/3
+         *
+         * Checking the user Id (owner) for the image and the logged in user id.
+         * If they are equal we allow the edit operation, else set the editError attribute
+         * and redirect to the same page
+         *
+         */
+
         Image image = imageService.getImage(imageId);
+        User user = (User) session.getAttribute("loggeduser");
 
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
+
+        if( image.getUser().getId()!=user.getId() ) {
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("editError", error);
+            model.addAttribute("tags", image.getTags());
+            return "images/image";
+        }
+
         return "images/edit";
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -153,7 +172,26 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+
+        /** PartA Issue #2: https://github.com/upgrad-edu/Course_4_Assignment/issues/3
+         *
+         * Checking the user Id (owner) for the image and the logged in user id.
+         * If they are equal we allow the edit operation, else set the editError attribute
+         * and redirect to the same page
+         *
+         */
+        Image image = imageService.getImage(imageId);
+        User user = (User) session.getAttribute("loggeduser");
+
+        if( image.getUser().getId()!=user.getId() ) {
+            model.addAttribute("image", image);
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError", error);
+            model.addAttribute("tags", image.getTags());
+            return "images/image";
+        }
+
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
